add price

+++ database/migrations/2021_06_26_123531_create_purchases_invoices_details_table.php
@@ -18,8 +18,8 @@ class CreatePurchasesInvoicesDetailsTable extends Migration
             $table->unsignedBigInteger('invoice_id');
             $table->string('product_name');
             $table->integer('quantity');
-            $table->integer('price');
-            $table->integer('total');
+            $table->double('price');
+            $table->double('total');
             $table->foreign('invoice_id')->references('id')->on('purchases_invoices')->onDelete('cascade');
             $table->timestamps();
         });
